migrated sidenav content to its own component
