+++ packages/client/src/app/app.component.html
@@ -1,84 +1,7 @@
 
 <mat-sidenav-container>
   <mat-sidenav [fixedInViewport]="true" #sidenav>
-    <div class="user-menu" *ngIf="currentUser; else login">
-      <div class="user-header">
-        <div class="avatar">
-          <img class="user-avatar" [src]="currentUser.profile.avatar">
-        </div>
-        <div class="username">
-          <h3>{{ currentUser.username }}</h3>
-          <h4>Welcome back!</h4>
-        </div>
-      </div>
-      <div class="user-buttons">
-        <mat-tab-group>
-          <mat-tab>
-            <ng-template mat-tab-label>
-              <i-feather name="compass"></i-feather>
-            </ng-template>
-            <ul>
-              <li class="item" *ngIf="checkUserRolesForContribMenu()"><a [routerLink]="['/dashboard']"><i-feather name="grid"></i-feather>Dashboard</a></li>
-              <li class="item"><a [routerLink]="['/portfolio', currentUser._id, currentUser.username | slugify]"><i-feather name="user"></i-feather>Portfolio</a></li>
-              <li class="item"><a [routerLink]="['/home/alerts']"><i-feather name="bell"></i-feather>Alerts</a></li>
-              <li class="item"><a [routerLink]="['/home/inbox']"><i-feather name="inbox"></i-feather>Inbox</a></li>
-              <li class="item"><a [routerLink]="['/home/settings']"><i-feather name="settings"></i-feather>Settings</a></li>
-              <li class="divider"></li>
-              <li class="item"><a (click)="logout()"><i-feather name="log-out"></i-feather>Log Out</a></li>
-            </ul>
-          </mat-tab>
-          <mat-tab>
-            <ng-template mat-tab-label>
-              <i-feather name="bell"></i-feather>
-            </ng-template>
-            for offprint notifications
-          </mat-tab>
-          <mat-tab>
-            <ng-template mat-tab-label>
-              <i-feather name="message-circle"></i-feather>
-            </ng-template>
-            for offprint convos
-          </mat-tab>
-          <mat-tab>
-            <ng-template mat-tab-label>
-              <i-feather name="layers"></i-feather>
-            </ng-template>
-            here's some history content
-          </mat-tab>
-        </mat-tab-group>
-      </div>
-      <div class="user-footer">
-        <p>Offprint Alpha, ver. 0.3.0</p>
-      </div>
-    </div>
-    <ng-template #login>
-      <div class="user-menu">
-        <div class="login-form">
-          <h2>Welcome to Offprint!</h2>
-          <form [formGroup]="loginForm" (ngSubmit)="onLoginSubmit()">
-  
-            <p>New around here? Why not <a [routerLink]="['/register']">register today</a>!</p>
-            
-            <input type="email" formControlName="email" class="u-full-width" placeholder="Email" autocomplete="email" required>
-  
-            <input type="password" formControlName="password" class="u-full-width" placeholder="Password" autocomplete="current-password" required>
-  
-            <label>
-              <input type="checkbox" formControlName="rememberMe">
-              <span class="label-body">Remember me for a month</span>
-            </label>
-  
-            <div class="loader" *ngIf="loadingLogin">
-              <div class="lds-dual-ring"></div>
-            </div>
-            <button type="submit" [disabled]="!loginForm.valid" class="button-primary" style="width: 100%;" *ngIf="!loadingLogin">Log in</button>
-          </form>
-        </div>
-        <div class="user-footer">
-          <p>Offprint Alpha, ver. 0.3.0</p>
-        </div>
-      </div>
-    </ng-template>
+    <site-sidenav></site-sidenav>
   </mat-sidenav>
   <mat-sidenav-content>
     <div class="nav">
