+++ packages/client/src/app/app.component.less
@@ -1,174 +1,3 @@
-/* Sidenav */
-.mat-tab-label-content {
-    i-feather {
-        stroke: whitesmoke;
-        stroke-width: 1;
-    }
-}
-
-.mat-tab-nav-bar, .mat-tab-header {
-    border-bottom: none !important;
-}
-
-div.user-menu {
-    display: flex;
-    flex-direction: column;
-    min-width: 35rem;
-    height: 100%;
-
-    div.login-form {
-        flex-grow: 1;
-        padding: 1.5rem;
-        h2 {
-            color: whitesmoke;
-            text-align: center;
-            border-bottom: 1px solid whitesmoke;
-        }
-    
-        form {
-            margin-bottom: 0;
-            input[type=email],input[type=password] {
-                box-shadow: none;
-                border: 1px solid whitesmoke;
-                color: whitesmoke;
-                background-color: var(--site-accent-hover);
-                &::placeholder {
-                    color: whitesmoke;
-                }
-            }
-            p {
-                text-align: center;
-                a {
-                    color: whitesmoke;
-                    text-decoration: none;
-                    background: linear-gradient(0deg, var(--site-accent) 0%, var(--site-accent-hover) 10%, var(--site-accent) 50%);
-                    &:hover {
-                        color: whitesmoke;
-                        background: linear-gradient(0deg, var(--site-accent) 0%, var(--site-accent-hover) 10%, var(--site-accent-hover) 90%, var(--site-accent) 100%);
-                        text-decoration: none;
-                    }
-                }
-            }
-
-            div.loader {
-                div.lds-dual-ring {
-                    &:after {
-                        border-color: whitesmoke transparent whitesmoke transparent;
-                    }
-                }
-            }
-        }
-    }
-
-    div.user-header {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        padding: 1.5rem;
-        height: auto;
-
-        div.avatar {
-            img.user-avatar {
-                max-width: 75px;
-                max-height: 75px;
-                object-fit: cover;
-                border: 1px solid whitesmoke;
-                border-radius: 50%;
-                font-size: 0;
-                margin-right: 1.5rem;
-            }
-        }
-
-        div.username {
-            font-size: 0;
-            flex: 1;
-            border-bottom: 1px solid whitesmoke;
-            h3 {
-                color: whitesmoke;
-                font-weight: 500;
-                font-size: 36px;
-                margin-bottom: 0;
-            }
-
-            h4 {
-                color: whitesmoke;
-                margin-top: 0;
-                margin-bottom: 0;
-                font-size: 24px;
-                font-style: italic;
-            }
-        }
-    }
-
-    div.user-buttons {
-        flex: 1;
-        width: 100%;
-        margin-top: 0.5rem;
-        
-        .mat-tab-group {
-            font-family: Raleway, sans-serif;
-            font-size: initial;
-            color: whitesmoke;
-        }
-
-
-
-        ul {
-            list-style-type: none;
-            padding: 1rem 1.5rem 0 1.5rem;
-            margin-bottom: 0;
-            li {
-                cursor: pointer;
-            }
-            li.item {
-                &:focus {
-                    border-radius: 0 !important;
-                    outline: none;
-                }
-                a {
-                    width: 100%;
-                    padding: 10px;
-                    font-size: 15px;
-                    color: whitesmoke;
-                    display: inline-block;
-                    box-sizing: border-box;
-                    &:hover {
-                        color: whitesmoke;
-                        background-color: var(--site-accent-hover);
-                    }
-                    i-feather {
-                        position: relative;
-                        width: 15px;
-                        height: 15px;
-                        top: 0.15rem;
-                        margin-right: 1rem;
-                    }
-
-                    &:focus {
-                        border-radius: 0 !important;
-                        outline: none;
-                    }
-                }
-            }
-            li.divider {
-                border-bottom: 1px solid whitesmoke;
-                margin-top: 0;
-                margin-bottom: 0.5rem;
-            }
-        }
-    }
-
-    div.user-footer {
-        border-top: 1px solid whitesmoke;
-        color: whitesmoke;
-        margin: 0 1.5rem 1.5rem 1.5rem;
-        p {
-            text-align: center;
-            margin: 0;
-        }
-    }
-}
-
 /* Navigation Bar */
 div.nav {
   width: 100%;
